+++ Classes/Foundation/Public/Helpers/CKStringHelper.m
@@ -32,8 +32,8 @@
                                                options:(1 << 0) //NSStringDrawingUsesLineFragmentOrigin
                                                context:nil];
     
-    CGFloat floorWidth  = floor(rect.size.width);
-    CGFloat floorHeight = floor(rect.size.height);
+    CGFloat floorWidth  = (CGFloat)(floor(rect.size.width));
+    CGFloat floorHeight = (CGFloat)(floor(rect.size.height));
 
     return CGSizeMake(floorWidth + ((floorWidth == rect.size.width) ? 0 : 1),floorHeight + ((floorHeight == rect.size.height) ? 0 : 1));
 }
