+++ Classes/UI/Public/Map/CKMapViewController.m
@@ -23,7 +23,7 @@
 
 
 CGFloat distance(MKMapPoint p1, MKMapPoint p2){
-    return sqrt(pow(p1.x-p2.x,2)+pow(p1.y-p2.y,2));
+    return (CGFloat)(sqrt(pow(p1.x-p2.x,2)+pow(p1.y-p2.y,2)));
 }
 
 NSInteger compareLocations(id <MKAnnotation>obj1, id <MKAnnotation> obj2, void *context)
